prov/efa: split rxr_op_entry.h from rxr.h
extract the definition of rxr_tx_entry and rxr_rx_entry to
a separate header file rxr_op_entry.h for further refactor.
