+++ prov/efa/Makefile.include
@@ -79,6 +79,7 @@ _efa_headers = \
 	prov/efa/src/rxr/rxr_pkt_cmd.h \
 	prov/efa/src/rxr/rxr_read.h \
 	prov/efa/src/rxr/rxr_atomic.h \
+	prov/efa/src/rxr/rxr_op_entry.h \
 	prov/efa/src/rxr/rdm_proto_v4.h
